requirements.txt
python-dotenv>=1.0.1
psycopg[binary]>=3.2.1
Flask>=3.0.0
